+++ src/api/Api.js
@@ -8,17 +8,18 @@ import {
   getRequestWithHeaderData,
 } from "./requestexecutor";
 
-import { REGISTER, 
-        LOGIN, 
-        USER_DATA, 
-        ASSESSMENT, 
-        GET_COUNSELLOR_PATIENTS, 
-        ALL_SELF_ASSESSMENT, 
-        GET_COUNSELLOR_DOCTORS, 
-        COUNSELLOR_GET_ALL_APPOINTMENT, 
-        DOCTOR_GET_ALL_APPOINTMENT,
-        PATIENT_GET_ALL_APPOINTMENT
-       } from "./url";
+import {
+  REGISTER,
+  LOGIN,
+  USER_DATA,
+  ASSESSMENT,
+  GET_COUNSELLOR_PATIENTS,
+  ALL_SELF_ASSESSMENT,
+  GET_COUNSELLOR_DOCTORS,
+  COUNSELLOR_GET_ALL_APPOINTMENT,
+  DOCTOR_GET_ALL_APPOINTMENT,
+  PATIENT_GET_ALL_APPOINTMENT,
+} from "./url";
 
 //------------------------------------------------ALL USERS---------------------------------------------------------------------------------
 export async function register(
@@ -55,9 +56,9 @@ export async function assesmentSubmission({ url = ASSESSMENT, data, token }) {
   return responseData;
 }
 
-export async function getSelfAssessment({url = ALL_SELF_ASSESSMENT, token, email}) {
-  url = url + "?email=" + email 
-  console.log("Request Body : " + url)
+export async function getSelfAssessment({ url = ALL_SELF_ASSESSMENT, token, email }) {
+  url = url + "?email=" + email;
+  console.log("Request Body : " + url);
   const responseData = await getRequestWithHeaderData(url, token);
   return responseData;
 }
@@ -82,6 +83,10 @@ export async function getCounsellorPatients(token) {
   return responseData;
 }
 
+export async function putCounsellorPatients(data, token) {
+  let responseData = await putRequest(GET_COUNSELLOR_PATIENTS, data, token);
+  return responseData;
+}
 
 export async function getCounsellorDoctors(token) {
   let responseData = await getRequestWithHeader(GET_COUNSELLOR_DOCTORS, token);
@@ -94,13 +99,14 @@ export async function counsellorGetAllAppointmentByDate({url=COUNSELLOR_GET_ALL_
   return responseData;
 }
 
+
 export async function doctorGetAllAppointmentByDate({url=DOCTOR_GET_ALL_APPOINTMENT, date,token}){
   url=url+"?appointment="+date;
   let responseData = await getRequestWithHeaderData(url,token);
   return responseData;
 }
 
-export async function patientGetAllAppointmentByDate(token){
-  let responseData = await getRequestWithHeader(PATIENT_GET_ALL_APPOINTMENT,token);
+export async function patientGetAllAppointmentByDate(token) {
+  let responseData = await getRequestWithHeader(PATIENT_GET_ALL_APPOINTMENT, token);
   return responseData;
 }