+++ src/sections/patients/appointment-table.js
@@ -1,21 +1,17 @@
 import PropTypes from "prop-types";
 
 import {
-  Avatar,
-  Box,
-  Card,
-  Checkbox,
-  Stack,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TablePagination,
   TableRow,
+  TableCell,
+  Button,
   Typography,
+  Card,
+  Container,
+  Box
 } from "@mui/material";
 import { Scrollbar } from "src/components/scrollbar";
 import { useRouter } from "next/navigation";
+
 export const AppointmentTable = (props) => {
   const {
     items = [],
@@ -26,27 +22,41 @@ export const AppointmentTable = (props) => {
   let count=0;
   return (  
 
-      <Scrollbar sx={{ mb: 3 }}>
+    <Box
+          component="main"
+          sx={{
+          flexGrow: 1,
+          py: 8
+        }}>
+        <Container maxWidth="xl">
+      
               {items.map((appointment) => {
                 count++;
                 if (appointment) {
                   return (
-                    <TableRow>
-
-                      <TableCell>{count}</TableCell>  
-                      <TableCell>Email :{appointment.Patient}</TableCell>
-                      <TableCell>Doctor : {appointment.Doctor}</TableCell>
-                      <TableCell>Counsellor : {appointment.Counsellor}</TableCell>
-                      <TableCell>Appointment : {appointment.Appointment}</TableCell>
-                      <TableCell>Assign Doctor : {appointment.AssignDoctor}</TableCell>
-                      <TableCell>{appointment.Desciption}</TableCell>
+                    
+                    <TableRow >
+                      <TableCell><Typography variant="h6">{count}</Typography></TableCell>  
+                      <TableCell><Typography variant="h6">Email :{appointment.Patient}</Typography></TableCell>
+                      <TableCell><Typography variant="h6">Doctor : {appointment.Doctor}</Typography></TableCell>
+                      <TableCell><Typography variant="h6">Counsellor : {appointment.Counsellor}</Typography></TableCell>
+                      <TableCell><Typography variant="h6">Appointment : {appointment.Appointment}</Typography></TableCell>
+                      <TableCell><Typography variant="h6">Assign Doctor : {appointment.AssignDoctor}</Typography></TableCell>
+                      <TableCell><Typography variant="h6">{appointment.Desciption}</Typography></TableCell>
+                      <TableCell> <Button onClick={() => handlegetAssessment(index, patient.Patient)} variant="contained" color="primary"> 
+                          Confirm
+                        </Button></TableCell>
+                      <TableCell> <Button onClick={() => handlegetAssessment(index, patient.Patient)} variant="contained" color="primary"> 
+                          Reject
+                        </Button></TableCell>
                     </TableRow>
+
                   );
                 }
               })}
-
-      </Scrollbar>
-
+        </Container>
+    </Box>
+             
   );
 };