+++ src/sections/patients/patients-table.js
@@ -9,19 +9,20 @@ import {
   TablePagination,
   Button,
   TableRow,
-  List, 
-  ListItem, 
-  ListItemText, 
+  List,
+  ListItem,
+  ListItemText,
   TextField,
   FormControl,
   InputLabel,
   Select,
   MenuItem,
-  Dialog, 
-  DialogTitle, 
+  Dialog,
+  DialogTitle,
   DialogContent,
   FormControlLabel,
-  Checkbox
+  Checkbox,
+  Typography,
 } from "@mui/material";
 
 import { Scrollbar } from "src/components/scrollbar";
@@ -29,6 +30,8 @@ import { getSelfAssessment } from "src/api/Api";
 import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
 import { convertToList } from "src/data/selfassessmentmap";
+import { putCounsellorPatients } from "src/api/Api";
+import { status } from "nprogress";
 
 export const PatientsTable = (props) => {
   const {
@@ -44,26 +47,31 @@ export const PatientsTable = (props) => {
     rowsPerPage = 0,
     selected = [],
   } = props;
-  
-  const [email, setEmail] = useState(""); 
-  const [openReport, setOpenReport] = useState(false); 
-  const [openAction, setOpenAction] = useState(false); 
-  const [listofQuestions, setList] = useState(new Map([])); 
 
-  const router = useRouter();
+  const [openReport, setOpenReport] = useState(false);
+  const [token, setToken] = useState("");
+  const [email, setEmail] = useState("");
+  const [id, setId] = useState("");
+  const [listofQuestions, setList] = useState(new Map([]));
 
-  const handlegetAssessment = async (index, email)=>{
-    const token = window.sessionStorage.getItem("token");
-    setEmail(email);
-    const res= await getSelfAssessment({token, email})
-    if(res){
-      setOpenReport(true);
-      const list = await convertToList(res);
-      setList(list)
-    }
-  }
+  const [openAction, setOpenAction] = useState(false);
+  const [selectedOption, setSelectedOption] = useState(null);
+  const [doctor, setDoctor] = useState(false);
+  const [date, setDate] = useState(null);
+  const [description, setDescription] = useState("");
+  const [isSaveDisabled, setIsSaveDisabled] = useState(false);
 
-  const itemTextStyle = {
+  const [errorMessageReport, setErrorMessageReport] = useState("");
+  const [errorMessageAction, setErrorMessageAction] = useState("");
+  const [successMessageAction, setSuccessMessageAction] = useState("");
+
+  const formattedDateTime = null;
+  const minDateTime = new Date().toISOString().slice(0, -8);
+  useEffect(() => {
+    setIsSaveDisabled(selectedOption === null);
+  }, [selectedOption]);
+
+  const styles = {
     primary: {
       fontSize: "16px",
       fontWeight: 600,
@@ -76,21 +84,105 @@ export const PatientsTable = (props) => {
     },
   };
 
-  const handleClickOpen = () => {
+  const router = useRouter();
+
+  const handleClickOpen = (email, id) => {
     setOpenAction(true);
+    setEmail(email);
+    setId(id);
+    console.log(window.sessionStorage.getItem("token"));
+    setToken(window.sessionStorage.getItem("token"));
   };
 
   const handleClose = () => {
+    console.log("Here");
     setOpenReport(false);
     setOpenAction(false);
   };
 
-  const [selectedOption, setSelectedOption] = useState(null);
-  const [status, setStatus] = useState(false);
+  const handlegetAssessment = async (email) => {
+    const token = window.sessionStorage.getItem("token");
+    console.log(email + " Email");
+    const res = await getSelfAssessment({ token, email });
+    if (res) {
+      setOpenReport(true);
+      const list = await convertToList(res);
+      setList(list);
+    }
+    //TODO ERROR
+  };
 
-  const handleReject = () => {
-    setSelectedOption(null);
-    setStatus(false);
+  const handleCheckboxChange = (event) => {
+    setSelectedOption(event.target.checked ? event.target.name : null);
+    setIsSaveDisabled(event.target.checked);
+    setErrorMessageAction("");
+    setSuccessMessageAction("");
+    setErrorMessageReport("");
+  };
+
+  const handleDateTimeChange = (event) => {
+    const newDateTime = new Date(event.target.value);
+    setDate(
+      newDateTime.toISOString().substring(0, 10) + "T" + newDateTime.toTimeString().substring(0, 8)
+    );
+    setErrorMessageAction("");
+    setSuccessMessageAction("");
+    setErrorMessageReport("");
+  };
+
+  const handleAction = async () => {
+    console.log(email + " Email");
+    console.log(id + " Id");
+    console.log(selectedOption);
+    if (
+      (selectedOption === "bookAppointment" && (description.length === 0 || !date)) ||
+      (selectedOption === "reject" && description.length === 0) ||
+      (selectedOption === "assign" && !doctor)
+    ) {
+      setErrorMessageAction("Fill the required information before saving");
+    } else {
+      if (selectedOption === "assign") {
+        //TODO: CALL ASSIGN DOCTOR API
+      } else {
+        let patientStatus = null;
+        let successMessage = "";
+        if (selectedOption === "reject") {
+          patientStatus = false;
+          successMessage = "Patient has been rejected";
+        } else {
+          patientStatus = true;
+          if (selectedOption === "bookAppointment") {
+            successMessage = "You have booked an appointment with the patient";
+          } else {
+            successMessage = "You have assign the patient to a doctor";
+          }
+        }
+
+        const data = {
+          id: id,
+          Patient: email,
+          Appointment: date,
+          Accept: patientStatus,
+          Description: description,
+        };
+
+        const response = await putCounsellorPatients(data, token);
+        console.log("Put Counselor Response1: " + response.id);
+        console.log("Put Counselor Response2: " + response.Error);
+        if (response.Error) {
+          setErrorMessageAction(response.Error);
+          setSuccessMessageAction("");
+        } else if (response.id || response.detail) {
+          setErrorMessageAction("");
+          setSuccessMessageAction(successMessage);
+        } else {
+          setErrorMessageAction(
+            "Action Failed. Something went wrong. Please contact the administrator"
+          );
+          setSuccessMessageAction("");
+        }
+      }
+    }
   };
 
   return (
@@ -111,130 +203,222 @@ export const PatientsTable = (props) => {
               {items.map((patient, index) => {
                 if (!patient.AssigntoDoctor) {
                   return (
-                    <TableRow key = {index}>
+                    <TableRow key={index}>
                       <TableCell>{patient.Firstname}</TableCell>
                       <TableCell>{patient.Lastname}</TableCell>
                       <TableCell>{patient.Patient}</TableCell>
                       <TableCell>
-                        <Button onClick={() => handlegetAssessment(index, patient.Patient)} variant="contained" color="primary"> 
+                        <Button
+                          onClick={() => handlegetAssessment(patient.Patient)}
+                          variant="contained"
+                          color="primary"
+                        >
                           Report
                         </Button>
-                        <Dialog             
-                          BackdropProps={{ invisible: false }}
+                        {errorMessageReport && (
+                          <Typography color="error" variant="body1">
+                            {errorMessageReport}
+                          </Typography>
+                        )}
+                        <Dialog
+                          BackdropProps={{
+                            invisible: true,
+                            style: { opacity: 1 },
+                          }}
+                          maxWidth="lg"
+                          fullWidth={true}
+                          PaperProps={{
+                            style: {
+                              maxHeight: "90vh",
+                              minHeight: "90vh",
+                            },
+                          }}
                           open={openReport}
                           onClose={handleClose}
-                          maxWidth="md"
-                          maxHeight="md">
+                        >
                           <DialogTitle>Self Assessment Result</DialogTitle>
                           <DialogContent>
                             <List>
                               {Array.from(listofQuestions).map(([question, answer], index) => (
                                 <ListItem key={index}>
-                                  <ListItemText 
-                                    primary={question} secondary={answer}      
-                                    primaryTypographyProps={{ style: itemTextStyle.primary }}
-                                    secondaryTypographyProps={{ style: itemTextStyle.secondary }}
+                                  <ListItemText
+                                    primary={question}
+                                    secondary={answer}
+                                    primaryTypographyProps={{ style: styles.primary }}
+                                    secondaryTypographyProps={{ style: styles.secondary }}
                                   />
                                 </ListItem>
                               ))}
                             </List>
                           </DialogContent>
-                        </Dialog>  
+                        </Dialog>
                       </TableCell>
                       <TableCell>
-                        <Button onClick={handleClickOpen}  variant="contained" color="primary">Action</Button>
-                        <Dialog open={openAction} onClose={handleClose} maxwidth="md" maxheight="md">
+                        <Button
+                          onClick={() => handleClickOpen(patient.Patient, patient.id)}
+                          variant="contained"
+                          color="primary"
+                        >
+                          Action
+                        </Button>
+                        <Dialog
+                          BackdropProps={{
+                            invisible: true,
+                            style: { opacity: 1 },
+                          }}
+                          maxWidth="sm"
+                          fullWidth={true}
+                          PaperProps={{
+                            style: {
+                              maxHeight: "80vh",
+                              minHeight: "40vh",
+                            },
+                          }}
+                          open={openAction}
+                          onClose={handleClose}
+                        >
                           <DialogTitle>Pick an action</DialogTitle>
                           <DialogContent>
-  <div>
-    <FormControlLabel
-      control={
-        <Checkbox
-          checked={selectedOption === "bookAppointment"}
-          onChange={(event) =>
-            setSelectedOption(
-              event.target.checked ? "bookAppointment" : null
-            )
-          }
-        />
-      }
-      label="Book appointment"
-      style={{ display: "block" }}
-    />
-    <FormControlLabel
-      control={
-        <Checkbox
-          checked={selectedOption === "accept"}
-          onChange={(event) =>
-            setSelectedOption(event.target.checked ? "accept" : null)
-          }
-        />
-      }
-      label="Accept"
-      style={{ display: "block" }}
-    />
-    <FormControlLabel
-      control={
-        <Checkbox
-          checked={selectedOption === "reject"}
-          onChange={(event) =>
-            setSelectedOption(event.target.checked ? "reject" : null)
-          }
-          onClick={() => handleReject()}
-         />
-      }
-      label="Reject"
-      style={{ display: "block" }}
-    />
-    {selectedOption === "bookAppointment" && (
-      <TextField
-        label="Date"
-        type="date"
-        onChange={(event) => setStatus(true)}
-      />
-    )}
-    {selectedOption === "accept" && (
-      <FormControl>
-        <InputLabel>Status</InputLabel>
-        <Select
-          value={status ? "Confirmed" : ""}
-          onChange={(event) =>
-            setStatus(event.target.value === "Confirmed")
-          }
-        >
-          <MenuItem value="Confirmed">Confirmed</MenuItem>
-          <MenuItem value="Pending">Pending</MenuItem>
-        </Select>
-      </FormControl>
-    )}
-
-  </div>
-  <Button sx={{
-    backgroundColor: '#4CAF50',
-    color: '#ffffff',
-    '&:hover': {
-      backgroundColor: '#357a38',
-    },
-    mr: 2, // adds margin-right of 16px (default spacing unit)
-  }}
-  onClick={() => handleClickOpen} variant="contained" color="primary" margintop="10"> 
-                          Save
-                        </Button>
-   <Button sx={{
-    backgroundColor: '#4CAF50',
-    color: '#ffffff',
-    '&:hover': {
-      backgroundColor: '#357a38',
-    },
-    mr: 2, // adds margin-right of 16px (default spacing unit)
-  }}
-  onClick={() => handleClose} variant="contained" color="primary"> 
-                          Cancel
-                        </Button>
-
+                            <div>
+                              <FormControlLabel
+                                control={
+                                  <Checkbox
+                                    name="bookAppointment"
+                                    checked={selectedOption === "bookAppointment"}
+                                    onChange={(event) => handleCheckboxChange(event)}
+                                  />
+                                }
+                                label="Book appointment"
+                                style={{ display: "block" }}
+                              />
+                              <FormControlLabel
+                                control={
+                                  <Checkbox
+                                    name="assign"
+                                    checked={selectedOption === "assign"}
+                                    onChange={(event) => handleCheckboxChange(event)}
+                                  />
+                                }
+                                label="Assign a doctor"
+                                style={{ display: "block" }}
+                              />
+                              <FormControlLabel
+                                control={
+                                  <Checkbox
+                                    name="reject"
+                                    checked={selectedOption === "reject"}
+                                    onChange={(event) => handleCheckboxChange(event)}
+                                  />
+                                }
+                                label="Reject"
+                                style={{ display: "block" }}
+                              />
+                              {selectedOption === "bookAppointment" && (
+                                <>
+                                  <TextField
+                                    variant="outlined"
+                                    size="small"
+                                    required
+                                    fullWidth
+                                    margin="normal"
+                                    InputLabelProps={{
+                                      shrink: true,
+                                      style: { marginTop: "0.2rem" },
+                                    }}
+                                    format="yyyy-MM-dd'T'HH:mm"
+                                    value={formattedDateTime}
+                                    inputProps={{ min: minDateTime }}
+                                    label="Date and Time"
+                                    type="datetime-local"
+                                    onChange={(event) => handleDateTimeChange(event)}
+                                  />
+                                  <TextField
+                                    variant="outlined"
+                                    size="small"
+                                    required
+                                    fullWidth
+                                    margin="none"
+                                    multiline
+                                    rows={4}
+                                    value={description}
+                                    label="Description"
+                                    type="description"
+                                    onChange={(event) => {
+                                      setDescription(event.target.value);
+                                      setErrorMessageAction("");
+                                      setErrorMessageReport("");
+                                      setSuccessMessageAction("");
+                                    }}
+                                  />
+                                </>
+                              )}
+                              {selectedOption === "assign" && (
+                                <FormControl>
+                                  <InputLabel>Doctor Name</InputLabel>
+                                  <Select
+                                    value={doctor ? "Confirmed" : ""}
+                                    onChange={(event) =>
+                                      setDoctor(event.target.value === "Confirmed")
+                                    }
+                                  >
+                                    <MenuItem value="Confirmed">Confirmed</MenuItem>
+                                    <MenuItem value="Pending">Pending</MenuItem>
+                                  </Select>
+                                </FormControl>
+                              )}
+                              {selectedOption === "reject" && (
+                                <>
+                                  <TextField
+                                    variant="outlined"
+                                    size="small"
+                                    required
+                                    fullWidth
+                                    margin="normal"
+                                    multiline
+                                    rows={4}
+                                    value={description}
+                                    label="Reason for Rejection"
+                                    type="description"
+                                    onChange={(event) => {
+                                      setDescription(event.target.value);
+                                      setErrorMessageReport("");
+                                      setSuccessMessageAction("");
+                                    }}
+                                  />
+                                </>
+                              )}
+                            </div>
+                            <Button
+                              sx={{
+                                backgroundColor: "#4CAF50",
+                                color: "#ffffff",
+                                "&:hover": {
+                                  backgroundColor: "#357a38",
+                                },
+                                mr: 2,
+                                mt: 1,
+                              }}
+                              disabled={isSaveDisabled}
+                              onClick={() => handleAction(event, patient.Patient)}
+                              variant="contained"
+                              color="primary"
+                              margintop="10"
+                            >
+                              Save
+                            </Button>
+                            {successMessageAction && (
+                              <Typography color="primary" variant="body1" fontWeight="bold">
+                                {successMessageAction}
+                              </Typography>
+                            )}
+                            {errorMessageAction && (
+                              <Typography color="error" variant="body1" fontWeight="bold">
+                                {errorMessageAction}
+                              </Typography>
+                            )}
                           </DialogContent>
-                        </Dialog>  
-                      </TableCell> 
+                        </Dialog>
+                      </TableCell>
                     </TableRow>
                   );
                 }